Create Action Types for Event State
