--- src/redux/events/eventsTypes.js
@@ -1,1 +1,0 @@
-// import * as types from "./actionTypes";
