+++ src/states/events/eventsTypes.js
@@ -1,0 +1,6 @@
+
+export const SUCCESS = 'SUCCESS';
+
+export const PENDING = 'PENDING';
+
+export const ERROR = 'ERROR';
